Remove reference to "client" certificate
Motivation
----------
Columnar doesn't use client certificates
(which are used for mutual TLS authentication).

This sentence was referring to the server certificate.
The user doesn't need to think about those, though.

This _might_ be a good place to mention that
if the Capella certificate authority is compromised,
you'll need to upgrade the SDK or configure it to use
a new certificate. However, that would require careful
wordsmithing, and can be added later.