Polish quickstart and installation guide
Motivation
----------
Minimize duplication.

Encourage people to use the example project
as a starting point.

Modifications
-------------
Promote "Maven Project Template" to a top-level
section of the quickstart guide. Include the SDK
version tag in the GitHub link to pin it to a
release version.

Renamed "Full Installation" -> "Maven Coordinates",
because the concept of full installation does not
apply to Java.

Instead of duplicating the maven coordinates in the
quickstart guide, link to the "Maven Coordinates" page.

Remove unused import for internal `Certificates` class.